v3.4.0: the big display face grows a third digit (reg 60 renders 0-999)
The STANDARD cabinets shelve more than 99 medicines, so reg 60's 0-99
face stopped fitting the job. Values under 100 keep the exact two-digit
look the pharmacists know ("45", not "045"); 100-999 grow a third digit
at the same glyph size, so nothing about the screen changes for the
numbers people already use.

The real constraint was never the panel, it was the bezel over the
glass: the 54 px font draws three digits 123/128 px wide, leaving 2.5 px
a side where the vertical crop already proved 5 px is needed. The font
is regenerated at --height 51 (round-trip check passes): three digits
now span 117 px with 5.5 px margins, and the bitmap shrinks ~300 B --
the build comes out 472 B SMALLER than v3.3.2 (59,888 B) despite the
new code, which matters with about 1 KB of flash left.

DEMO's counter widens with it: a board with ID 108 now shows "108"
instead of the % 100 remainder "08", and taps count to 999.

Contract note: reg 60 was published to the hospital's server team as
0-99. The clamp moving to 999 is backward-compatible for writers that
stayed in range, but the pick-sequence document must say so when this
ships. No settings/stats schema change -- OTA from any v3.3.x is safe.

+++ src/drivers/oled.h
@@ -29,11 +29,13 @@ void oledClear();
  */
 void oledPrint(const char *text, uint8_t textSize = 2);
 
-/*  @brief Show a two-digit number centered using a large bitmap font.
+/*  @brief Show a number (up to three digits) centered using a large bitmap
+ *         font. Values below 100 keep the familiar two-digit face ("05");
+ *         100-999 grow a third digit at the same glyph size.
  *
- *  @param value Number to display. Values above 99 are wrapped by the caller.
+ *  @param value Number to display, clamped to 999.
  */
-void oledPrintLargeNumber(uint8_t value);
+void oledPrintLargeNumber(uint16_t value);
 
 /*  @brief Show a small centered title with a large centered number below it,
  *         using the standard GFX font (no bitmap digits).
